// app/(routes)/tax-break/_utils/type.ts
export type ITaxBreak = {
  createdBy: string;
  taxBreaksId: string;
  taxBreaksCode: string;
  name: string;
  documentRequired: string;
  eligilityCriteria: string;
  applicableTo: string;
  type: string;
  updatedBy: null | string;
  breaksType: null | string;
  uuid: string;
  createdAt: string;
  deletedAt: null | string;
  updatedAt: string;
};